requirements.txt
Django
djangorestframework
redis
pymongo
django-cors-headers
pandas